+++ src/constant.rs
@@ -1,6 +1,6 @@
-pub const PI: f64 = 3.14159265358979323846264338327950288_f64;
-pub const EPS: f64 = 1e-5;
-pub const INF: f64 = 1e5;
+pub const PI: f32 = 3.14159265358979323846264338327950288_f32;
+pub const EPS: f32 = 1e-5;
+pub const INF: f32 = 1e5;
 
 // pub const WIDTH: usize = 320 * 1;
 // pub const HEIGHT: usize = 240 * 1;
@@ -23,8 +23,8 @@ pub const SPP: usize = 64;
 // pub const HEIGHT: usize = 1080;
 // pub const SPP: usize = 10000;
 
-pub const TIME_LIMIT: f64 = 4.0 * 60.0 + 28.0;
-pub const SAVE_IMAGE_INTERVAL: f64 = 10.0;
-// pub const TIME_LIMIT: f64 = 5.0;
-// pub const SAVE_IMAGE_INTERVAL: f64 = 5.0;
-pub const SAVE_IMAGE_INTERVAL_ERROR: f64 = 1.0;
+pub const TIME_LIMIT: f32 = 4.0 * 60.0 + 28.0;
+pub const SAVE_IMAGE_INTERVAL: f32 = 10.0;
+// pub const TIME_LIMIT: f32 = 5.0;
+// pub const SAVE_IMAGE_INTERVAL: f32 = 5.0;
+pub const SAVE_IMAGE_INTERVAL_ERROR: f32 = 1.0;
